SourceCode/requirements.txt
appdirs==1.4.4
attrs==20.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
cloud-init==19.1.6
command-not-found==0.3
configobj==5.0.6
cytoolz==0.11.0
dbus-python==1.2.16
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
filelock==3.0.12
Flask==1.1.2
gmpy2==2.0.8
gmssl==3.2.1
hexbytes==0.2.1
idna==2.8
ipfsapi==0.4.4
ipfshttpclient==0.7.0a1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
language-selector==0.1
lru-dict==1.1.7
MarkupSafe==1.1.1
multiaddr==0.0.9
netaddr==0.8.0
netifaces==0.10.4
oauthlib==3.1.0
parsimonious==0.8.1
pbr==5.5.1
protobuf==3.16.0
pycryptodome==3.10.1
PyGObject==3.36.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.1
PyYAML==5.3.1
requests==2.24.0
requests-unixsocket==0.2.0
rlp==2.0.1
six==1.15.0
ssh-import-id==5.10
stevedore==3.3.0
toolz==0.11.1
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
varint==1.0.2
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
web3==5.19.0
websockets==8.1
Werkzeug==1.0.1
